inference/requirements.txt
grpcio==1.64.1
Pillow==10.3.0
protobuf==5.27.1
torch==2.3.1
transformers==4.41.2
